+++ ext/session/serializer_user.c
